chore(deps): upgrade dotenv to v10
Upgrade the `dotenv` package to v10.0.0 or above. This didn't seem to
cause any issues in the tests, and I'm running on the `.nvmrc`-defined
version of Node v10.x. Since you wanted to drop support for Node 8 & 9,
I don't see any reason to test it further back in time.

Resolves #65